#hard coded shader point light #redid a lot of the core rendernig code
#todo, more descriptor set work... like a lot
#set up deferred rendering
#set up multi light system to use with deferred rendering
#gpu compute shaders for noise gen

+++ VulkanTest/MeshBufferHandler.cpp
@@ -1,6 +1,6 @@
 #include "MeshBufferHandler.h"
 
-MeshBufferHandler::MeshBufferHandler(VmaAllocator allocator, std::vector<Vertex> vertices, std::vector< uint16_t> indices) :
+MeshBufferHandler::MeshBufferHandler(VmaAllocator allocator, std::vector<Vertex> vertices, std::vector< uint32_t> indices) :
 	vertices(vertices),
 	indices(indices)
 {
